Show average turns with decimal places, fix Vietnamese bronze apple translation, upgrade Gradle to 8.2.0

+++ scripts/src/main/java/io/github/fate_grand_automata/scripts/entrypoints/AutoBattle.kt
@@ -148,7 +148,7 @@ class AutoBattle @Inject constructor(
         val averageTimePerRun: Duration,
         val minTurnsPerRun: Int,
         val maxTurnsPerRun: Int,
-        val averageTurnsPerRun: Int
+        val averageTurnsPerRun: Double
     )
 
     private fun makeExitState(): ExitState {
